Adding a Lazy Data Transfer Object
A DtoLazy is able to store a reference to Dto and Load itself
lazily. This is needed in a case we need to make relation to Dto
in another DtoCollection, but that DtoCollection was not fetched
from refresh parser and saved to DB.

As a part of the saving, we want to turn LazyDto object to an
ActiveRecord object hidden in the Dto object's object method.