Add the InterRegionApiMethodRelay mixin
This will allow any method on a model to be easily routed to a remote
region using a new DSL defined by the mixin.

When the method is called it will be first filtered through logic
to determine if the subject of the method belongs to the current
region. If it does, the call will continue to the model logic, but
if it does not, a REST API call will be issued for the method in question.

The blocks that accompany the DSL can be used to deliver API post
payload to the relay method based on the parameters of the original
method (which will be passed to the block).

In the case of class methods, the block is required and must yield
two values, the target resource id (in order to determine the region
the call should be relayed to), and then the API post payload (or nil).