Don't expand actions in action results, expand for bulk queries
I did a bad refactoring back in
https://github.com/ManageIQ/manageiq/pull/11512. In that PR, I reasoned
that in `#get_aspecs` (see
[here](https://github.com/ManageIQ/manageiq/blob/c08abf4b3ded3c082e65301634aa86d4f555a26d/app/controllers/api_controller/renderer.rb#L110-L129)),
the controller would never respond to `typed_actions` (unless one such
method was created in the future), so this path could be removed. What I
failed to see was some implicit knowledge about how actions are
rendered:

1. `typed_actions` comes from the `:resource_actions` option, which is
set in `#render_collection_type`.
2. The *presence* of that argument decided whether `#get_aspecs` should
return anything.
3. Action results don't originate from `#render_collection_type`, since
they are a type of `POST` (`#render_collection_type` is called from
`#show`).
4. Hence, action results shouldn't expand actions.

The solution tries to make that knowledge more explicit by adding an
`:expand_actions` option, defaulting to false and set to `true` for
GETs, which should restore the original behavior.