Do not store :cloud_subnet in data when saved as nested
Do not store :cloud_subnet in data when saved as nested. We
allowed cloud_subnets to be saved as a toplevel object, so we
should list it in data only then. Otherwise it's enough to
have them in data_index for crosslinking.

TBD. only AWS fetches cloud subnets in one API query, other
providers can be optimized like this.