Set @options only once
There are three possible callers of build_results_for_report_chargeback
each of them sets @options to the same value that is later passed in as
an argument.

Let's have a single place to set it.

In future I want to get a rid of class variable. But it will take few
moments before we get there.